Create menuPrincipal.ts
Remove menuPrincipal logic from main
